fixed mentioned code smells in the review
