+++ .gitignore
@@ -1,0 +1,2 @@
+/Backend-Internship-main/PHASE-03-CodeReview/my-app/.idea/
+/Backend-Internship-main/PHASE-03-CodeReview/my-app/target/
